fix(project): Increase the typescript linting/formatting for more consistent code
The current method of sharing JavaScript/Typescript linting rules is
non-standard, I've now updated it to included the linting rules from the
DSP3 project. It may be that we need to fork that DSP3 library for one
specific to this app, but that needs discussion as to what and why.

IssueID EODHP-614