feat(accessibility): Add highlighting of accessibility issues while developing
I've added the Axe Core dependency and added a custom component to
manage adding this only when not in production and only if a browser
is being used (not in tests etc). This will write accessibility issues
found to the Browser's console along with the severity of the issue.

IssueID EODHP-619