feat(tests): Add Network layer mocking
I've added MSWJS, this uses a Service Worker to intercept network
requests before they hit the internet. Mocking the Network layer
is better than mocking actual code as you have more confidence in
your code, since it isn't mocked, the Network layer is not in our
control, the request is actually sent, just intercepted. This allows
us to write the Network requests as we would normally, intercept them
and return any data we see fit. This is especially useful for mocking
error responses and mocking out services like paywalls, which would
be called if we allowed the network request to complete, during
testing this could become very expensive.

IssueID EODHP-618