fix(project): Format all JavaScript/TypeScript files
Now that we have better linting and introduced the `prettier` formatter,
I have now re-formatted each TypeScript file by opening and saving
it. This proves the integration with VS Code is properly setup. This
has raised a couple of issues with clickable divs etc which is bad
for accessibility issues, but we should deal with them separately.

IssueID OEDHP-614