fix(map): Add simple drawing tool
I've added a simple bounding box drawing tool, rather than use the
map extent for searches. It might still not be obvious to users that
they need to draw an AOI first before they can see scenes, so it might
be better to put the button as the first filter in the sidebar, need
to discuss this with others.

IssueID EODHP-620