fix(filters): Disable QA placeholder option
The placeholder option of the dropdown/select should not be selectable.
It is only there to inform the user to select something, what we don't
want is someone actually selecting it and sending a dummy value.
To prevent this I've disabled the option, so that it is still displayed
but once the user selects a real option, they cannot then go back
and select this placeholder.

IssueID EODHP-639