feat(frontend): Integrate Purchase forms into app
We have agreed to treat selecting a collection item as a commercial
item for now. This will change once a designer is in place and a
proper UI designed. Once the user has selected an item, then the
UI changes to show the accordions for the query and purchase forms.

IssueID EODHP-629