fix(frontend): Make draggable panel only drag from header
The previous incarnation of the panel allowed the user to click
and drag the panel from anywhere inside it. However, this blocked
the `slider` in the query form from working, so I updated the
draggable to only drag when you select the header, much like many
desktop apps do. I re-styled the header as well, so as to make it
clearer to the user.

IssueID EODHP-629