feat(frontend): Lazily load components
I have lazily loaded some components that are not required for the
app to render. These components are generally required when the user
clicks something and new UI elements are displayed. This reduces the
size of the chunks being download, so that it is as small as possible
to speed up the initial render of the app.

IssueID EODHP-792