Revert "feat(frontend): Lazily load components"
This reverts commit 9b0ca0dcf873ba2d50b3fda98c10890134d1385f.

I am reverting this commit as the lazy loading is breaking the app
when deployed to `dev`. I don't know exactly why, all I know is it
gets 404 errors when requesting the lazy loaded code. This is less
important than the rest of the changes though, so shouldn't stop
that from being merged.