fix(frontend): Fix bootstrap bleed-through CSS
As has been noted, bootstrap is bleeding through to the react app
but is only visible when the app is deployed. I have therefore added
some extra styling to undo what bootstrap is doing for this particular
situation, but we need a more holistic approach to this problem.

IssueID EODHP-629