fix(map): Refactor map drawing tools
I previously added a very basic box drawing tool as it was simpler
than trying to use the map bounds. I have now added more complex
drawing tools for irregular polygons and circles.

I also took to the opportunity to change the filter as before it
was called `bounds` which is an actual term for the rectangular area
encompassing a polygon. Since we are not using bounds, but the drawn
Area of Interest (AOI), I felt it better to rename this.

IssueID EODHP-689