feat(project): Lazy load components
I've lazy loaded a few components and this is working during development
but when it was deployed to the `dev` environment it broke. I'm not yet
sure what the cause is but it is likely to do with file reving and
un-reving going on in the github action and how the app is then linked
into the web-presence.

IssueID EODHP-792