feat(frontend): Use state in URL for collection/item API calls
User can come from another part of the system and configure the
catalogue based on their private catalogue. This means the URL
needs to contain state to configure the UI.

I've updated the app to read the URL and use the `path` state if
it exists in requests for collections/items.

IssueID EODHP-811