feat(toolbox): Display and copy to clipboard collection item URLs
Before we wanted to show the purchase forms when a user clicked on
a collection item. This has changed now and there is a button on
the item for this. Now when a user clicks the item we show an `assets`
panel that displays what assets there are, the URL and a button to
copy the URL to the clipboard, so users can use it elsewhere.

IssueID EODHP-663