fix(toolbox): Fix draggable toolbox in strict mode
The react-draggable library used to provide the capability to drag
the toolbox has a warning when using JavaScript `Strict mode`. The
warning itself leads to documentation on this. In fact it is an
easy fix, instead of having the library find the DOM node, we tell
it by using a `ref` hook.