feat(frontend): Redirect to sign in on Forbidden response
Now that we can configure the Resource catalogue with the path to a
private catalog, it is possible to get a 403 Forbidden response. This
means the user is not logged in (this should change to a 401 Unauthorized
but that is a problem for another repo).

If we get a Forbidden response, then we redirect to the sign in page,
so the user can sign in. Once that happens, they should be redirected
back to the catalogue, with the name of the private catalog still intact.
Now, if all that is try, the user can then see the catalog of interest.

IssueID EODHP-811