fix(toolbox): Fix merging mistake
I merged the `main` branch in but removed this code as it looked
wrong. I now see it is more a type guard to ensure the StacItem
is a Polygon and therefore has `geometry coordinates`. I'm not sure
just silently returning from the function is the right solution but
it was code that already existed.