feat(toolbox): Add skeleton when loading collection items
I've refactored the code to display collection items so that we better
inform the user what the state is. For instance, it is possible for a
user to filter the collections, but not select an AOI, so the user is
informed, I have extended this to now show a skeleton of the items when
the app has the AOI and is fetching the items, when the data is retrieved
then the skeleton is replaced with the actual items. This prevents a
lot of re-drawing on the UI which is better UX and gives the user an
idea of what the UI will look like once the data is received.

IssueID EODHP-834