fix(accessibility): Fix linting issues for accessibility
Previously when adding linting rules, we found a number of accessibility
issues, like being able to click on static elements. Normally you click
on buttons/links, not divs or spans. This isn't bad, but it is recommend
to use non-static elements where possible. I have now changed clickable
divs for buttons and styled the buttons accordingly, so they match what
the div looked like.

I did remove one bit of code where we open a collection item link in a
new window, as we now have the copy-to-clipboard feature. The code I
removed was a hidden feature anyway as you had to press the `shift`
button and click the link, but there is no docs anywhere to say this,
so most would never know it exists. The copy-to-clipboard is a far
better way of doing this and gives the user the option of what to
do with the URL e.g. open in a new tab, another app, another part of
this app etc.

IssueID EODHP-836