fix(toolbox): Refactor getting of the STAC catalog URL
I previously simplified the extraction of the `catalogUrl` but failed
to understand what the original URL could look like. During review it
was pointed out that there could be sub catalogs. There needs to be
some discussion around this as we don't believe it is predictable at
this point in time, so I have reverted the code back to what it was
with one minor change. Instead of just replacing the env Var, it can
now be an absolute or relative URL, so I calculate the index of the
end of it and remove everything up to that point, this is no different
from what it was before.