WIP fix(layout): Fix layout so it grows to take up space
The Resource catalogue does not stand alone, rather it sits inside
a django webpage. It has been requested that there be no scrollbar
on the app, so we need a way for the center area i.e. the map and
sidebar, to grow to whatever size is available. This means changing
code in the web-presence and in the resource catalogue. The web-presence
we set the height of the header/footer, so that the size the middle
section (sidebar and map) can be calculated.

I've added some styling here so that the sidebar/map can grow to
take all the available space left.

IssueID EODHP-484