fix(Data Catalogue): Fix strange pagination results
When going back and forth in the pagination for the Data Catalogue
we saw that sometimes there were 6 results, others 9, some seemed
duplicates etc.

The cause was that the table is a list of rows, but some rows had
the exact same `key`. This is a common error when mapping over an
array of items, ensuring the key is unique fixed the issue.

I think the real cause is maybe bad data, but I'd need to chat with
those that know the data better. We need to ensure I feel that there
can be no 2 bits of data with the exact same `id`, or what good is it.

IssueID EODHP-844