fix(toolbox): Display toolbox in Safari properly
The cause of this issue was the styling specifying a webkit specific
setting, even though webkit is the engine used by safari. Removing it
fixed the problem for me during triage, it maybe worked at some point
but maybe other styling since then broke it. Its easy to happen if we
don't invest in testing against multiple browsers in an automated way.

IssueID EODHP-816