ISO8601 Implementation.
Altered the DataPoint type to have a value attribute instead of text, this allows us to render react components as the DataPoint's output giving us more control over the style of the value. For a temporal DataPoint, instead of rendering just the interval, as a string, it is now being rendered as a component, this is because the styling was inline and overflowing which was visually unpleasant. Gave the option to include the time as a string at the end of the formatted date.