Addressed review comments and tidied up code.
Extracted common variables and exported them where needed. Removed the need to include the parameter includeTime and insted put this in the switch case statement for when the user requests that particular formatting style. Gave the type DataPoint an ID based on the collection is it a part of and the field type that it is representing.