fix(toolbox): Sort spatial filtering on back link
If you have filtered catalogs/collections and select a collection,
then go back to the treeview, we lost the nodes bbox. I have added
a guard clause to only filter by spatial bbox if there actuall is a
bbox to filter by ie. not just an empty array.

IssueID EODHP-981