918 ISO8601
* Initial time util implementation.

Added a new time util to handle basic string / date conversion to ISO8601.

* Added tests.

Added new date.tests.ts file. Added 5 tests to ensure that the method is working as expected.

* ISO8601 Implementation.

Altered the DataPoint type to have a value attribute instead of text, this allows us to render react components as the DataPoint's output giving us more control over the style of the value. For a temporal DataPoint, instead of rendering just the interval, as a string, it is now being rendered as a component, this is because the styling was inline and overflowing which was visually unpleasant. Gave the option to include the time as a string at the end of the formatted date.

* Fixed test issues.

Test failing on throwing error. Fixed the method to throw error correctly and fixed test.

* Added react-datepicker.

The filter sidebar temporal input was using HTML input type date. The issue with this is that the format of the date is dependant on the user's locale. I have installed react-datepicker and configured it to use ISO8601 format.

* Fixed css linting error.

* Addressed review comments and tidied up code.

Extracted common variables and exported them where needed. Removed the need to include the parameter includeTime and insted put this in the switch case statement for when the user requests that particular formatting style. Gave the type DataPoint an ID based on the collection is it a part of and the field type that it is representing.

* Properly used exported date format in datepicker input.

* Made format method more generic.

Removed hard coded switch statement allowing for more flexibility.

* Removed redundant tests which no longer apply.

* Removed console.log