Added URL params for map and filter
Made a method on the FilterContext which can add a param to the URL and not refresh the page when added. Using this method on the inputs in the filter and also added event listeners to the map which listen for a change in centre and zoom. When the centre or zoom of the map is changed this is added to the URL.