934 date management (#38)
* Added date handling for collections and items.

Added a method to go through the attributes of collections and items and extract the first non null date and set this to be the rendered date value. Further discussion may be needed as to decide which date to use if multiple are present.

* Fixed linting errors.

* Added priority when chosing dates.

The date chosen is based off all available dates on a STAC object. In addition to this, the dates on the assets are no longer used when considering the dates available on an object.

* Updated method name for clarity.

Method name before did not adequately describe what the method returned. Remaned for clarity.

---------

Co-authored-by: Mark Small <mark.small@astrosat.space>