Can read in params from the URL using queries.
Added the ability to add queries to the URL and then read them in correctly into the application. Moved the setup for the VectorSource and VectorLayer of the AOI onto the map context as I feel this makes more sense. Moved the reading of the queries into the App context as it requires access to both the Filter and Map context.