URL follows API structure.
I am now using the correct catalogue path and using this in the URL. When the user reloads the page, the currently selected catalogue is being used when fetching the root catalogue. This is inline with the new design being created. In addition to this, the current view either map list or qa is also stored in the path.