1009 integrate ol stack (#44)
* Added preview.

Added a preview button to each item. When clicked adds a STACLayer to the map and moves the map there. Everything seems to be working as intended but I am getting some serious performance issues, more testing required plus 1007 may fix the performance issues.

* Linting.

* Only single preview at a time and overlay.

User can only view one preview at a time, if the user selects another preview then the currently shown one is removed from the map. Added a red button overlay to the top right of the preview which will remove the layer and the overlay itself from the map.

* Linting.

* Added ol-stac to package.json.

* Fixed styling.

Made button the same size as the other buttons.

* Type now detected and used in layer.

Preview will not show unless type specificed. I added a method to try and detect the type and this seems to have fixed the issue.