1014 dataset details page (#47)
* Added new DataDetails page.

Added a new icon to a collection that when pressed will send the user to a page dedicated to that collection. The DataDetails page gets the collection from the URL.

* Added initial page.

Rendering basic information about the collection. Formatting to follow the radiant earth stac browser. Including an OL map to show the items associated to the collection/

* Can access from all pages and urlk updates correctly.

* Styling and fetching improvements.

Makes a request for the collection based off the URL, this will make it easier to extend for items when that's needed. Improved the styling, though this is subject to change when we get the new design documents.

* Fixed styling and tests.

* Linting

* Addressed review comments.